requirements.txt
boto3
Pillow
python-resize-image
python-dotenv
celery
